Changed Array::New length from int32_t to size_t

+++ Coral.Native/Include/Coral/Array.hpp
@@ -8,13 +8,13 @@ namespace Coral {
 	class Array
 	{
 	public:
-		static Array New(int32_t InLength)
+		static Array New(size_t InLength)
 		{
 			Array<TValue> result;
 			if (InLength > 0)
 			{
 				result.m_Ptr = static_cast<TValue*>(Memory::AllocHGlobal(InLength * sizeof(TValue)));
-				result.m_Length = InLength;
+				result.m_Length = static_cast<int32_t>(InLength);
 			}
 			return result;
 		}
